fix(ci): satisfy clippy -D warnings; WIP-alpha README notice
CI gates clippy with -D warnings; the v0.0.4-alpha commit tripped several
default lints. Resolve them (behavior-preserving):

- edit.rs: collapse nested if-let into a let-chain; drop unused lifetime
- history/pg: replace redundant closures with bare fn paths
- proxy/handlers: collapse nested PII-detection if-lets into let-chains
- ui: collapse if-lets (auth/mcp/webhooks), unnecessary cast, map_or ->
  is_some_and, manual modulo -> is_multiple_of, allow too_many_arguments
- remove now-dead pages/coming.rs (all surfaces have real pages)

Also add a WIP/alpha notice to the top of the README.

+++ crates/drgtw-ui/src/pages/config.rs
@@ -279,7 +279,7 @@ fn render_page_inner(
                             (bool_field("traces", "traces", cfg.otel.traces))
                             (bool_field("metrics", "metrics", cfg.otel.metrics))
                             (bool_field("metrics_include_key_id", "metrics_include_key_id", cfg.otel.metrics_include_key_id))
-                            (float_field("sample_ratio", "sample_ratio", cfg.otel.sample_ratio as f64))
+                            (float_field("sample_ratio", "sample_ratio", cfg.otel.sample_ratio))
                         }
                     }
                 }
@@ -327,6 +327,7 @@ fn render_page_inner(
 // Section form wrapper
 // ---------------------------------------------------------------------------
 
+#[allow(clippy::too_many_arguments)]
 fn section_form(
     _state: &UiState,
     section: &str,
